+++ internal/cmd/restore.go
@@ -99,7 +99,6 @@ func runRestore(c *cli.Context) error {
 		return errors.Wrap(err, "init configuration")
 	}
 
-	db.LoadConfigs()
 	db.SetEngine()
 
 	// Database